feat(useMergeReducer-hook): convert mergeReducer function into a full reusable hook
