--- src/hooks/reducers/mergeReducer.js
@@ -1,10 +1,0 @@
-export default function mergeReducer (state, fragment) {
-  const resolvedFragment = typeof fragment === 'function'
-    ? fragment(state)
-    : (fragment ?? {})
-
-  return {
-    ...(state ?? {}),
-    ...resolvedFragment,
-  }
-}
